+++ buffer/buffer_stubborn2.c
@@ -1,8 +1,8 @@
 #include <errno.h>
 #include "buffer.h"
 
-int buffer_stubborn_read(ssize_t (*op)(),int fd,const char* buf, size_t len,void* cookie) {
-  int w;
+ssize_t buffer_stubborn_read(ssize_t (*op)(),int fd,const char* buf, size_t len,void* cookie) {
+  ssize_t w;
   for (;;) {
     if ((w=op(fd,buf,len,cookie))<0)
       if (errno == EINTR) continue;
