+++ dns.h
@@ -57,7 +57,7 @@ unsigned int dns_domain_length(const char *);
 int dns_domain_equal(const char *,const char *);
 int dns_domain_suffix(const char *,const char *);
 unsigned int dns_domain_suffixpos(const char *,const char *);
-int dns_domain_fromdot(char **,const char *,unsigned int);
+int dns_domain_fromdot(char **,const char *,size_t);
 int dns_domain_todot_cat(stralloc *,const char *);
 
 unsigned int dns_packet_copy(const char *,unsigned int,unsigned int,char *,unsigned int);
