+++ ip6.h
@@ -9,16 +9,16 @@
 extern "C" {
 #endif
 
-unsigned int scan_ip6(const char* src,char* ip);
-unsigned int fmt_ip6(char* dest,const char* ip);
-unsigned int fmt_ip6c(char* dest,const char* ip);
+size_t scan_ip6(const char* src,char* ip);
+size_t fmt_ip6(char* dest,const char* ip);
+size_t fmt_ip6c(char* dest,const char* ip);
 
-unsigned int scan_ip6if(const char* src,char* ip,uint32* scope_id);
-unsigned int fmt_ip6if(char* dest,const char* ip,uint32 scope_id);
-unsigned int fmt_ip6ifc(char* dest,const char* ip,uint32 scope_id);
+size_t scan_ip6if(const char* src,char* ip,uint32* scope_id);
+size_t fmt_ip6if(char* dest,const char* ip,uint32 scope_id);
+size_t fmt_ip6ifc(char* dest,const char* ip,uint32 scope_id);
 
-unsigned int scan_ip6_flat(const char *src,char *);
-unsigned int fmt_ip6_flat(char *dest,const char *);
+size_t scan_ip6_flat(const char *src,char *);
+size_t fmt_ip6_flat(char *dest,const char *);
 
 /*
  ip6 address syntax: (h = hex digit), no leading '0' required
