+++ readclose.h
@@ -1,14 +1,17 @@
 #ifndef READCLOSE_H
 #define READCLOSE_H
 
+/* for ssize_t */
+#include <sys/types.h>
+
 #include "stralloc.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
-int readclose_append(int fd,stralloc *buf,size_t initlen);
-int readclose(int fd,stralloc *buf,size_t initlen);
+ssize_t readclose_append(int fd,stralloc *buf,size_t initlen);
+ssize_t readclose(int fd,stralloc *buf,size_t initlen);
 
 #ifdef __cplusplus
 }
